requirements.txt
# Dependencies for Encoder.py and Decoder.py
Pillow
bitarray
